# fix(plugins): Don't copy test list during collection