@marioevz 2025-07-24
@felix314159 I reworked the function to make it safer while still skipping copying the list.

I timed the benchmark test collection before and after this commit:
### Before
```
(ethereum-execution-spec-tests) ➜  execution-spec-tests git:(update-all-benchmark-tests) time uv run fill --clean --fork=Prague --evm-bin=evmone-t8n --gas-benchmark-values 1,10,30,45,60,90,120 -m benchmark --collect-only > /dev/null
uv run fill --clean --fork=Prague --evm-bin=evmone-t8n --gas-benchmark-values  185.64s user 0.77s system 99% cpu 3:06.42 total
```

Time: 3:06.42

### After
```
(ethereum-execution-spec-tests) ➜  execution-spec-tests git:(update-all-benchmark-tests) time uv run fill --clean --fork=Prague --evm-bin=evmone-t8n --gas-benchmark-values 1,10,30,45,60,90,120 -m benchmark --collect-only > /dev/null
uv run fill --clean --fork=Prague --evm-bin=evmone-t8n --gas-benchmark-values  27.25s user 0.70s system 99% cpu 27.953 total
```
Time: 27.953